Moves getContrastYIQ to utils folder
